entity 분리 - CartList, AddedItems, ApplyCoupon, OrderSummary
