+++ src/refactoring/components/CartList/CartList.tsx
@@ -1,10 +1,11 @@
-// src/refactoring/components/CartList/CartList.tsx
 import { getAppliedDiscount } from "../../models/cart"
 import { Coupon } from '../../../types';
 import { useAtom } from "jotai";
 import { cartAtom } from "../../atoms/cartAtoms";
 import { useCart } from "../../hooks/useCart";
-
+import { AddedItems } from "./AddedItems";
+import { ApplyCoupon } from "./ApplyCoupon";
+import { OrderSummary } from "./OrderSummary";
 interface Props {
   coupons: Coupon[];
 }
@@ -17,80 +18,24 @@ export const CartList = ({ coupons }: Props) => {
   return (
     <div>
       <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
-      <div className="space-y-2">
-        {cart.map(item => {
-          const appliedDiscount = getAppliedDiscount(item);
-          return (
-            <div key={item.product.id} className="flex justify-between items-center bg-white p-3 rounded shadow">
-              <div>
-                <span className="font-semibold">{item.product.name}</span>
-                <br/>
-                <span className="text-sm text-gray-600">
-                  {item.product.price}원 x {item.quantity}
-                  {appliedDiscount > 0 && (
-                    <span className="text-green-600 ml-1">
-                      ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
-                    </span>
-                  )}
-                </span>
-              </div>
-              <div>
-                <button
-                  onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                  className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                >
-                  -
-                </button>
-                <button
-                  onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                  className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
-                >
-                  +
-                </button>
-                <button
-                  onClick={() => removeFromCart(item.product.id)}
-                  className="bg-red-500 text-white px-2 py-1 rounded hover:bg-red-600"
-                >
-                  삭제
-                </button>
-              </div>
-            </div>
-          );
-        })}
-      </div>
+      <AddedItems
+        cart={cart}
+        getAppliedDiscount={getAppliedDiscount}
+        removeFromCart={removeFromCart}
+        updateQuantity={updateQuantity}
+      />
 
-      <div className="mt-6 bg-white p-4 rounded shadow">
-        <h2 className="text-2xl font-semibold mb-2">쿠폰 적용</h2>
-        <select
-          value={selectedCoupon ? coupons.findIndex(c => c.code === selectedCoupon.code) : ""}
-          onChange={(e) => applyCoupon(coupons[parseInt(e.target.value)])}
-          className="w-full p-2 border rounded mb-2"
-        >
-          <option value="">쿠폰 선택</option>
-          {coupons.map((coupon, index) => (
-            <option key={coupon.code} value={index}>
-              {coupon.name} - {coupon.discountType === 'amount' ? `${coupon.discountValue}원` : `${coupon.discountValue}%`}
-            </option>
-          ))}
-        </select>
-        {selectedCoupon && (
-          <p className="text-green-600">
-            적용된 쿠폰: {selectedCoupon.name}
-            ({selectedCoupon.discountType === 'amount' ? `${selectedCoupon.discountValue}원` : `${selectedCoupon.discountValue}%`} 할인)
-          </p>
-        )}
-      </div>
+      <ApplyCoupon
+        coupons={coupons}
+        applyCoupon={applyCoupon}
+        selectedCoupon={selectedCoupon}
+      />
 
-      <div className="mt-6 bg-white p-4 rounded shadow">
-        <h2 className="text-2xl font-semibold mb-2">주문 요약</h2>
-        <div className="space-y-1">
-          <p>상품 금액: {totalBeforeDiscount.toLocaleString()}원</p>
-          <p className="text-green-600">할인 금액: {totalDiscount.toLocaleString()}원</p>
-          <p className="text-xl font-bold">
-            최종 결제 금액: {totalAfterDiscount.toLocaleString()}원
-          </p>
-        </div>
-      </div>
+      <OrderSummary
+        totalBeforeDiscount={totalBeforeDiscount}
+        totalAfterDiscount={totalAfterDiscount}
+        totalDiscount={totalDiscount}
+      />
     </div>
   )
 }
